requirements.txt
Pillow>=10.0.0
esptool>=4.0.0


